refactor: use common stylesheet for notification with button component
