--- frontend/src/components/molecules/AlertWithTwoButtons/style.css
@@ -1,17 +1,0 @@
-.datapass-container-two-btn {
-  display: flex;
-  gap: 1rem;
-  margin-top: 2rem;
-}
-
-.datapass-container-two-btn > .fr-alert {
-  flex-grow: 3;
-  margin-top: 0;
-}
-
-.datapass-two-btn {
-  display: flex;
-  flex-direction: column;
-  gap: 1rem;
-  align-self: center;
-}
