--- frontend/src/components/molecules/HighlightWithButton/style.css
@@ -1,15 +1,0 @@
-.datapass-container {
-  display: flex;
-  gap: 1rem;
-  align-items: center;
-  justify-content: space-between;
-}
-
-.datapass-container .datapass-highlight {
-  flex-grow: 1;
-  flex-basis: 0;
-}
-
-.datapass-span {
-  margin-bottom: 1.5rem;
-}
