+++ frontend/src/components/molecules/notification-with-buttons/HighlightWithButton.js
@@ -3,13 +3,12 @@ import Button from '../../atoms/hyperTexts/Button';
 import './style.css';
 
 export const HighlightWithButton = ({ title, children, onClick, label }) => (
-  <div className="fr-container--fluid fr-mt-5w datapass-container">
-    <Highlight className="datapass-highlight">
+  <div className="datapass-notification-with-buttons">
+    <Highlight>
       <span className="fr-h5 fr-mb-3w">{title}</span>
-      <br className="datapass-span" />
-      {children}
+      <p>{children}</p>
     </Highlight>
-    <div className="fr-container--fluid datapass-container">
+    <div className="datapass-notification-buttons">
       <div>
         <Button outline icon="edit" onClick={onClick}>
           {label}
