+++ frontend/src/components/organisms/form-sections/HeadSection/CallToProcessedMessageNotification.js
@@ -1,6 +1,6 @@
 import { useContext } from 'react';
 import { markEventsAsProcessed } from '../../../../services/enrollments';
-import AlertWithTwoButtons from '../../../molecules/AlertWithTwoButtons';
+import AlertWithTwoButtons from '../../../molecules/notification-with-buttons/AlertWithTwoButtons';
 import { FormContext } from '../../../templates/Form';
 import { OpenMessagePromptContext } from '../../../templates/Form/OpenMessagePromptContextProvider';
 import useListItemNavigation from '../../../templates/hooks/use-list-item-navigation';
